+++ CHANGELOG.md
@@ -3,7 +3,8 @@
  - Put SHTCX to sleep after reads
  - Poll data ready before reading TMP117 in single-shot mode
  - Fix TMP117 
- - Add sink-only GPIO CFG options for I2C. 
+ - Add sink-only GPIO CFG options for I2C.
+ - Add RD_WARNING_DEPRECATED error code
 
 ## 3.8.0
  - Store boot counter in flash